requirements.txt
Django==5.0.7
django-filter==24.2
whitenoise==6.6.0
